requirements.txt
flask-restful
flask-sqlalchemy
python-dotenv
flask-marshmallow
marshmallow-sqlalchemy